Add files via upload
Single-player rock/paper/scissors game. Both the player and the computer draw 3 cards from a shared deck or a randomized number of rocks, papers, and scissors. The player will select a card from their hand, and the computer will randomly select one of their own. Based on the cards played the player will win, lose, or tie. Clicking the lose/tie/win button will redraw cards for the player and the computer. Once the deck is depleted a certain amount, the deck will be remade.